CU-870 Remove mixin classes and add kupfer_commons mixins
kupfer-commons [1] contains the mixins `OrganizationCreateMixin` and
`OrganizationQuerySetMixin`, so the library is being imported here and
the related code deleted.

To install the private library, a `GITHUB_TOKEN` must be added to the
request made on `pip install`. For the local environment, the developers'
token must be added on `.env`. For deployed environments, the token is
gotten from drone secrets [2].

[1] https://github.com/kupfer-software/kupfer_commons/
[2] kupfer-software/kupfer_timetracking_service#65